+++ libs/eval/tests/test_task.cpp
@@ -44,7 +44,7 @@ TEST(TaskTest, PrepareAndRun) {
 
   // Initially not prepared
   EXPECT_FALSE(task.prepared);
-  EXPECT_EQ(task.run_count, 0);
+  EXPECT_EQ(task.run_count, static_cast<std::size_t>(0));
 
   // Prepare task
   task.prepare(data);
@@ -53,11 +53,11 @@ TEST(TaskTest, PrepareAndRun) {
 
   // Run task
   task.run(5);
-  EXPECT_EQ(task.run_count, 5);
+  EXPECT_EQ(task.run_count, static_cast<std::size_t>(5));
 
   // Run again to accumulate
   task.run(3);
-  EXPECT_EQ(task.run_count, 8);
+  EXPECT_EQ(task.run_count, static_cast<std::size_t>(8));
 }
 
 TEST(TaskTest, Reset) {
@@ -68,13 +68,13 @@ TEST(TaskTest, Reset) {
   task.run(10);
 
   EXPECT_TRUE(task.prepared);
-  EXPECT_EQ(task.run_count, 10);
+  EXPECT_EQ(task.run_count, static_cast<std::size_t>(10));
   EXPECT_EQ(task.input_name, "input1");
 
   // Reset should clear all state
   task.reset();
   EXPECT_FALSE(task.prepared);
-  EXPECT_EQ(task.run_count, 0);
+  EXPECT_EQ(task.run_count, static_cast<std::size_t>(0));
   EXPECT_TRUE(task.input_name.empty());
 }
 
@@ -86,7 +86,7 @@ TEST(TaskTest, SharedPointer) {
   tasks.push_back(task);
   tasks.push_back(std::make_shared<DummyTask>("task2"));
 
-  EXPECT_EQ(tasks.size(), 2);
+  EXPECT_EQ(tasks.size(), static_cast<size_t>(2));
   EXPECT_EQ(tasks[0]->name, "shared_task");
   EXPECT_EQ(tasks[1]->name, "task2");
 }
